build: move esbuild default macros into separate file and create ESM macro
This commit moves the existing ESBuild macros from `defaults.bzl` into a
dedicated file. As part of this commit we add a new macro called
`esbuild_esm_bundle`. This macro will output `.mjs` files and setup the
CJS/ESM interop automatically.

There is no need in requesting the prodmode output since it builds on
the devmode `.d.ts` and therefore requires two actions. Devmode and
Prodmode are equivalent output-wise so we should always use devmode as
if we already combined demvode & prodmode.