build: enable spec bundling for jasmine tests to work with ESM
This is somewhat similar to how we run ESM tests in the components
repository. We now bundle all specs at build time into a single ESM
file. This ensures the tests run as expected (because the Bazel NodeJS
rules do not fully support ESM yet). Also helps us validating the tool
in a way we also ship it to our users (i.e. bundled).

As part of this change, we also wire up the bootstrap jasmine code that
registers an `beforeEach` call. This keeps the behavior similar to the
jasmine bootstrap logic we had before.

Lastly, this commit resets the ng-dev config between tests as we will
need to use `setConfig` more heavily with ESM (where we cannot simply
spy on the `getConfig` module-exported member).